requirements.txt
streamlit~=1.37
pandas~=2.2
plotly~=5.23
pydantic~=2.7
python-dotenv~=1.0
requests~=2.32
msal~=1.29
tenacity~=9.0
email-validator~=2.2
SQLAlchemy~=2.0
pyodbc~=5.2
azure-identity~=1.17
azure-keyvault-secrets~=4.8
opentelemetry-sdk~=1.26
opentelemetry-exporter-otlp~=1.26
ruff~=0.6
black~=24.8
mypy~=1.11
pytest~=8.3
